api/requirements.txt
neo4j==5.25.0 
langchain==0.3.1
langchain-community==0.3.1
langchain-experimental==0.3.2
langchain-google-genai==2.0.1
langchain-groq==0.2.0
python-dotenv==1.0.1
tiktoken==0.7.0
networkx==3.3
streamlit==1.38.0
matplotlib==3.9.2
seaborn==0.13.2
pymongo==4.9.1
scikit-learn==1.5.2
statsmodels==0.14.3
xgboost==2.1.1
prophet==1.1.5
plotly==5.24.1
pyvis==0.3.2
openpyxl==3.1.5
langgraph==0.2.38
streamlit_folium==0.23.1
fastapi==0.115.3
uvicorn==0.32.0











